+++ src/mantle/support/class-collection.php
@@ -88,7 +88,7 @@ class Collection implements ArrayAccess, Enumerable {
 	 * @param  (callable(int): TTimesValue)|null $callback
 	 * @return static<int, TTimesValue>
 	 */
-	public static function times( $number, callable $callback = null ) {
+	public static function times( $number, ?callable $callback = null ) {
 		if ( $number < 1 ) {
 			return new static();
 		}
@@ -412,7 +412,7 @@ class Collection implements ArrayAccess, Enumerable {
 	 * @param (callable(TValue, TKey): bool)|null $callback
 	 * @return static
 	 */
-	public function filter( callable $callback = null ) {
+	public function filter( ?callable $callback = null ) {
 		if ( $callback ) {
 			return new static( Arr::where( $this->items, $callback ) );
 		}
@@ -429,7 +429,7 @@ class Collection implements ArrayAccess, Enumerable {
 	 * @param  TFirstDefault|(\Closure(): TFirstDefault) $default
 	 * @return TValue|TFirstDefault
 	 */
-	public function first( callable $callback = null, $default = null ) {
+	public function first( ?callable $callback = null, $default = null ) {
 		return Arr::first( $this->items, $callback, $default );
 	}
 
@@ -707,7 +707,7 @@ class Collection implements ArrayAccess, Enumerable {
 	 * @param  TLastDefault|(\Closure(): TLastDefault) $default
 	 * @return TValue|TLastDefault
 	 */
-	public function last( callable $callback = null, $default = null ) {
+	public function last( ?callable $callback = null, $default = null ) {
 		return Arr::last( $this->items, $callback, $default );
 	}