+++ src/mantle/testing/concerns/trait-hooks.php
@@ -23,7 +23,7 @@ trait Hooks {
 	 */
 	public function hooks_set_up(): void {
 		if ( ! self::$hooks_saved ) {
-			$this->backup_hooks();
+			self::backup_hooks();
 		}
 	}
 
@@ -31,7 +31,7 @@ trait Hooks {
 	 * Routines to run during tearDown().
 	 */
 	public function hooks_tear_down(): void {
-		$this->restore_hooks();
+		self::restore_hooks();
 	}
 
 	/**
@@ -44,9 +44,8 @@ trait Hooks {
 	 * @global array $wp_current_filter
 	 * @global array $wp_filter
 	 */
-	protected function backup_hooks() {
-		$globals = [ 'wp_actions', 'wp_current_filter' ];
-		foreach ( $globals as $global ) {
+	protected static function backup_hooks(): void {
+		foreach ( [ 'wp_actions', 'wp_current_filter' ] as $global ) {
 			self::$hooks_saved[ $global ] = $GLOBALS[ $global ];
 		}
 		self::$hooks_saved['wp_filter'] = [];
@@ -63,10 +62,9 @@ trait Hooks {
 	 * @global array $wp_current_filter
 	 * @global array $wp_filter
 	 */
-	protected function restore_hooks() {
+	protected static function restore_hooks(): void {
 		// phpcs:disable WordPress.WP.GlobalVariablesOverride,WordPress.NamingConventions.PrefixAllGlobals
-		$globals = [ 'wp_actions', 'wp_current_filter' ];
-		foreach ( $globals as $global ) {
+		foreach ( [ 'wp_actions', 'wp_current_filter' ] as $global ) {
 			if ( isset( self::$hooks_saved[ $global ] ) ) {
 				$GLOBALS[ $global ] = self::$hooks_saved[ $global ];
 			}