+++ src/mantle/testing/mail/class-mock-mailer.php
@@ -15,7 +15,7 @@ if ( ! file_exists( ABSPATH . '/wp-includes/PHPMailer/PHPMailer.php' ) ) {
 	return;
 }
 
-require_once ABSPATH . '/wp-includes/PHPMailer/PHPMailer.php';
+require_once ABSPATH . '/wp-includes/PHPMailer/PHPMailer.php'; // @phpstan-ignore-line requireOnce.fileNotFound
 
 /**
  * Mock PHPMailer class.