+++ tests/Testing/Concerns/MakesHttpRequestsTest.php
@@ -30,8 +30,11 @@ class MakesHttpRequestsTest extends Framework_Test_Case {
 		putenv( 'MANTLE_EXPERIMENTAL_TESTING_USE_HOME_URL_HOST=' );
 
 		remove_all_actions( 'template_redirect' );
+<<<<<<< HEAD
 
 		update_option( 'home', 'http://' . WP_TESTS_DOMAIN );
+=======
+>>>>>>> origin/1.x
 	}
 
 	public function test_get_home() {
@@ -41,7 +44,7 @@ class MakesHttpRequestsTest extends Framework_Test_Case {
 
 	public function test_get_singular() {
 		$post_id = static::factory()->post->create();
-		$this->get( get_permalink( $post_id ) )
+		$this->get( get_permalink( $post_id ) . '?test=asda' )
 			->assertQueryTrue( 'is_single', 'is_singular' )
 			->assertQueriedObjectId( $post_id )
 			->assertSee( get_the_title( $post_id ) );
@@ -76,9 +79,10 @@ class MakesHttpRequestsTest extends Framework_Test_Case {
 	public function test_get_term() {
 		$category_id = static::factory()->category->create();
 
-		$this->get( get_term_link( $category_id, 'category' ) );
-		$this->assertQueryTrue( 'is_archive', 'is_category' );
-		$this->assertQueriedObjectId( $category_id );
+		$this
+			->get( get_term_link( $category_id, 'category' ) )
+			->assertQueryTrue( 'is_archive', 'is_category' )
+			->assertQueriedObjectId( $category_id );
 	}
 
 	public function test_wordpress_404() {
@@ -148,7 +152,7 @@ class MakesHttpRequestsTest extends Framework_Test_Case {
 			}
 		);
 
-		$this->post( '/test-post' )
+		$this->post( '/test-post', [ 'request-param' => 'bar' ], [ 'test-header' => 'test-value' ] )
 			->assertCreated()
 			->assertHeader( 'test-header', 'test-value' )
 			->assertContent( 'yes' );