+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+📢 Minimum PHP version is now 8.2. The framework supports 8.2 - 8.4.
+
 ### Added
 
 - ✨ Experimental feature ✨: Use the home URL as the base URL for testing rather
@@ -19,10 +21,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	the hard-coded `WP_TESTS_DOMAIN`. It will also infer the HTTPS status from
 	the home URL.
 - Added `with_option()`/`with_home_url()`/`with_site_url()` methods to the installation manager.
+- Add a `without_local_object_cache()` method to prevent the `object-cache.php` drop-in from being loaded locally.
+- Added a better `dump()` method to the response object when testing HTTP
+  requests that will dump the request/response to the console.
 
 ### Changed
 
+- Removed support for PHP 8.1. The minimum PHP version is now 8.2.
+- For projects that require PHPUnit 9, the `phpunit/phpunit` version is now set to `^9.6.22`.
+- Upgraded to Symfony 7.0 packages.
 - Disable `spatie/once`'s cache if found during unit testing.
+- Ensure that the `QUERY_STRING` server variable is set when testing HTTP
+	requests.
+
+### Fixed
+
+- Ensure that built-in taxonomies properly register their rewrite rules during testing.
 
 ## v1.3.3 - 2025-01-10